requirements.txt
aiohttp==3.9.0
beautifulsoup4==4.12.2
flet==0.17.0
jieba==0.42.1
matplotlib==3.8.0
Pillow==10.0.1
qrcode==7.4.2
Requests==2.31.0
selenium==4.16.0
wordcloud==1.9.2
